+++ examples/ar_summarize.php
@@ -19,7 +19,6 @@
 <?php
 
 error_reporting(E_STRICT);
-$time_start = microtime(true);
 
 /*
   // Autoload files using Composer autoload
@@ -153,13 +152,6 @@ END;
 ENDALL;
 
 highlight_string($code);
-
-$time_end = microtime(true);
-$time = $time_end - $time_start;
-
-echo "<hr />Total execution time is $time seconds<br />\n";
-echo 'Amount of memory allocated to this script is ' . memory_get_usage() . ' bytes';
-
 ?>
 </div>
 </body>