+++ examples/mktime.php
@@ -15,7 +15,6 @@
 <?php
 
 error_reporting(E_STRICT);
-$time_start = microtime(true);
 
 date_default_timezone_set('UTC');
 
@@ -60,13 +59,6 @@ $code = <<< END
 END;
 
 highlight_string($code);
-
-$time_end = microtime(true);
-$time = $time_end - $time_start;
-
-echo "<hr />Total execution time is $time seconds<br />\n";
-echo 'Amount of memory allocated to this script is ' . memory_get_usage() . ' bytes';
-
 ?>
 </div>
 </body>